refactor: SideNavBar 컴포넌트 서버/클라이언트 분리 (#13)
- ThemeToggle 컴포넌트 분리 및 클라이언트 컴포넌트로 전환
- SideNavBar를 서버 컴포넌트로 변경
- 성능 최적화 및 관심사 분리를 위한 구조 개선

------------------------------------------------------

------------------------------------------------------
