+++ components/common/Switch/index.tsx
@@ -1,0 +1,13 @@
+'use client'
+
+import * as styles from './styles.css'
+
+const Switch = () => {
+    return (
+        <button className={styles.themeToggle} onClick={() => {}}>
+            <span className={styles.themeIndicator}></span>
+        </button>
+    )
+}
+
+export default Switch
